Rename: format.ts -> index.utils.ts 및 getThemeMode 함수 이동
